feat: migrate PBMAgeUtils class to swift
