+++ PrebidMobile/PrebidMobileRendering/AdTypes/AdView/PBMWebView.m
@@ -34,7 +34,6 @@
 #import "PBMTouchDownRecognizer.h"
 #import "PBMViewExposure.h"
 #import "PBMCreativeViewabilityTracker.h"
-#import "PBMAgeUtils.h"
 #import "PBMWKScriptMessageHandlerLeakAvoider.h"
 #import "UIView+PBMExtensions.h"
