+++ PrebidMobileTests/PrebidMobileTest-Bridging-Header.h
@@ -101,7 +101,6 @@
 #import "PBMAdRequesterVAST.h"
 #import "PBMCreativeModelCollectionMakerVAST.h"
 #import "PBMTransaction.h"
-#import "PBMAgeUtils.h"
 #import "PBMVideoCreative.h"
 #import "PBMVideoView.h"
 #import "PBMVideoViewDelegate.h"
